LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/constructor/007-expected.txt

CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0000b'

PASS WebSockets: new WebSocket(url, null char)
